requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
channels==4.2.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.1.6
django-environ==0.12.0
djangorestframework==3.15.2
idna==3.10
kombu==5.4.2
prompt_toolkit==3.0.50
psycopg==3.2.4
psycopg-binary==3.2.4
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
